% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_release.R
\name{adapt_zenodo}
\alias{adapt_zenodo}
\title{Adapt the \code{.zenodo.json} file from \code{protocolsource} to \code{protocols} repository}
\usage{
adapt_zenodo(json = ".zenodo.json", write = TRUE, path_from = ".", path_to)
}
\arguments{
\item{json}{a \code{JSON} string, url or file}

\item{write}{whether to write a \code{.zenodo.json} file (TRUE) or to return a
\code{json} string}

\item{path_from}{Default current working directory. The root folder of the
\code{protocolsource} repo.}

\item{path_to}{Path to where the file should be written.}
}
\description{
Adapt the \code{.zenodo.json} file from \code{protocolsource} to \code{protocols} repository
}
\keyword{internal}
