% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_captions.R
\name{add_captions}
\alias{add_captions}
\title{Touch up figure and table captions after using \code{convert_docx_to_rmd()}}
\usage{
add_captions(
  from,
  to,
  name_figure_from = "Figuur",
  name_table_from = "Tabel",
  name_figure_to = "Figuur",
  name_table_to = "Tabel"
)
}
\arguments{
\item{from}{The \code{.Rmd} file to convert.
Can be given as an absolute or relative path.}

\item{to}{The filename to write the resulting \code{.Rmd} file.
Can be given as an absolute or relative path.}

\item{name_figure_from}{name that is given to figures in captions and cross
references in the \code{.Rmd} provided in \code{from} (default is \code{Figuur})}

\item{name_table_from}{name that is given to tables in captions and cross
references in the \code{.Rmd} provided in \code{from} (default is \code{Tabel})}

\item{name_figure_to}{name that should be given to figures in cross
references in the output \code{.Rmd} (default is \code{Figuur})}

\item{name_table_to}{name that should be given to tables in cross
references in the output \code{.Rmd} (default is \code{Tabel})}
}
\description{
After using \code{convert_docx_to_rmd()}, figure and table captions
are part of the main text and cross references to them are just text and
numbers, making it difficult to integrate these \code{.Rmd} files in reports:
they are not automatically renumbered, these captions cannot be listed,...
This function replaces the figure and table descriptions with true captions,
and cross references to these figures and tables with dynamic references
(that refer to the figure or table label).
}
\details{
This function expects an input that is generated by
\code{convert_docx_to_rmd()}, with \itemize{
\item  figure captions below the figure and starting with \code{Figuur }
(or other name to be defined in variable \code{name_figure}),
then a unique number and finally the description in one paragraph
\item  table captions above the table and starting with \code{Tabel }
(or other name to be defined in variable \code{name_table}),
then a unique number and finally the description in one paragraph
\item figures consisting of one image, not 2 figures near each other
\item cross references having the same label as the figure or table they
refer to, default \code{Figuur } or \code{Tabel } followed by the unique number
written in exactly the same way
}
Please check carefully for mistakes after using this function
}
\seealso{
Other convert:
\code{\link[=convert_docx_to_rmd]{convert_docx_to_rmd()}}
}
\concept{convert}
