% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dependencies.R
\name{add_dependencies}
\alias{add_dependencies}
\title{Adds dependencies to the YAML of an \code{index.Rmd} file}
\usage{
add_dependencies(
  code_mainprotocol,
  protocol_code,
  version_number,
  params,
  appendix = !is.na(params)
)
}
\arguments{
\item{code_mainprotocol}{Protocol code of the protocol for which dependencies
need to be declared in the YAML of its \code{index.Rmd} file}

\item{protocol_code}{Character vector of protocol codes that are dependencies
to the main protocol.}

\item{version_number}{Character vector of version numbers corresponding with
protocol_code.}

\item{params}{List of lists with protocol-specific parameters corresponding
with parameters from the protocols in protocol_code. Use \code{NA} if no
parameters should be set for a protocol.}

\item{appendix}{Logical vector indicating whether or not a dependency needs
to be included as a subprotocol (at the end of the main protocol in an
appendix). Default is \code{!is.na(params)}. When \code{params} is not \code{NA}, \code{appendix}
will always be set to \code{TRUE} even if the user passes another value.}
}
\description{
Adds dependencies to the YAML of an \code{index.Rmd} file
}
\examples{
\dontrun{
protocolhelper::add_dependencies(
  code_mainprotocol = "spp-999-en",
  protocol_code = c("sfp-123-en", "spp-124-en"),
  version_number = c("2020.01", "2020.02"),
  params = list(NA, list(width = 8, height = 8))
)
}
}
\seealso{
Other creation:
\code{\link[=add_one_subprotocol]{add_one_subprotocol()}},
\code{\link[=add_subprotocols]{add_subprotocols()}},
\code{\link[=create_protocol]{create_protocol()}},
\code{\link[=insert_protocolsection]{insert_protocolsection()}},
\code{\link[=update_protocol]{update_protocol()}},
\code{\link[=update_version_number]{update_version_number()}}
}
\concept{creation}
