% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_subprotocol.R
\name{add_subprotocol}
\alias{add_subprotocol}
\title{Function to add (a section of) a published protocol as a subprotocol
to a standard project protocol.}
\usage{
add_subprotocol(
  code_subprotocol,
  version_number,
  file_name,
  section = NULL,
  demote_header = c(0, 1, 2, -1),
  params = NULL,
  fetch_remote = TRUE
)
}
\arguments{
\item{code_subprotocol}{Character string giving the protocol code from
which a subprotocol will be made (usually a sfp-type protocol)}

\item{version_number}{Character string with format YYYY.NN}

\item{file_name}{Character string with the name of the Rmarkdown
file (a chapter starting with a level 1 heading).}

\item{section}{Optional character string with the name of a section within an
Rmarkdown file.
Can also be a unique substring of a section title.
If not specified (the default): the whole file is taken. It
is assumed that the section has a level 2 heading.}

\item{demote_header}{Number of '#' to prefix to all titles before inserting
in current protocol. Default is 0.
A negative value can be used to remove
'#' from all section titles.
Allowed values are visible in the usage section.}

\item{params}{A list of parameter name-value pairs in case you need to use
non-default values in parameterized protocols.}

\item{fetch_remote}{Whether or not to fetch the remote. Default TRUE.}
}
\value{
The function will return Rmarkdown
}
\description{
The idea is to execute this function in an R chunk with knitr
option \code{results="asis"}.
}
\examples{
\dontrun{
add_subprotocol(
  code_subprotocol = "sfp-401-nl",
  version_number = "2021.01",
  file_name = "07_stappenplan.Rmd",
  params = list(
     shape = "square",
     width = "3",
     height = "3"
  )
)
}
}
