% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_subprotocols.R
\name{add_subprotocols}
\alias{add_subprotocols}
\title{Render all sub-protocols belonging to a main protocol to single
markdown files}
\usage{
add_subprotocols(code_mainprotocol, fetch_remote = TRUE)
}
\arguments{
\item{code_mainprotocol}{Character string giving the protocol code for the
main protocol}

\item{fetch_remote}{Whether or not to fetch the remote. Default TRUE.}
}
\description{
The function should be called interactively (in the console)
after the dependencies section in the \code{YAML} header of the \code{index.Rmd} file
of the main protocol has been filled in with the aid of the
\code{protocolhelper::add_dependencies()} function.
For reproducibility, it is good practice to save the call in a
separate R script.
For each sub-protocol a single markdown file and associated media and
data files will be written.
Each sub-protocol will be written to a subfolder of the main
protocol.
The subfolder name is the same as the version number of the
sub-protocol.
}
\seealso{
Other creation:
\code{\link[=add_dependencies]{add_dependencies()}},
\code{\link[=add_one_subprotocol]{add_one_subprotocol()}},
\code{\link[=create_protocol]{create_protocol()}},
\code{\link[=insert_protocolsection]{insert_protocolsection()}},
\code{\link[=update_protocol]{update_protocol()}},
\code{\link[=update_version_number]{update_version_number()}}
}
\concept{creation}
