% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all.R
\name{check_all}
\alias{check_all}
\title{Check protocol \code{frontmatter} and \code{structure}}
\usage{
check_all(protocol_code, fail = !interactive())
}
\arguments{
\item{protocol_code}{Character string giving the protocol code}

\item{fail}{Should the function drop an error in case of a problem?
Defaults to \code{TRUE} in a non-interactive session and \code{FALSE} in an interactive
session.}
}
\value{
A report of all failed checks.
}
\description{
Combines \code{check_frontmatter()} and \code{check_structure()} in one function.
}
\seealso{
Other check:
\code{\link[=check_all_person_info]{check_all_person_info()}},
\code{\link[=check_frontmatter]{check_frontmatter()}},
\code{\link[=check_structure]{check_structure()}},
\code{\link{protocolcheck}},
\code{\link[=validate_orcid]{validate_orcid()}}
}
\concept{check}
