% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_all_author_info}
\alias{check_all_author_info}
\title{Helper function to check if author information is correct}
\usage{
check_all_author_info(author_list, problems_vect)
}
\arguments{
\item{author_list}{the yaml front matter part containing author info}

\item{problems_vect}{character vector of previously encountered problems}
}
\value{
a character vector of previously encountered problems and problems
identified for author names and \code{orcid} ids.
}
\description{
Checks the format of author names and \code{orcid} ids.
}
\seealso{
Other check: 
\code{\link{check_all}()},
\code{\link{check_frontmatter}()},
\code{\link{check_structure}()},
\code{\link{protocolcheck}},
\code{\link{validate_orcid}()}
}
\concept{check}
