% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_all_person_info}
\alias{check_all_person_info}
\title{Helper function to check if person information is correct}
\usage{
check_all_person_info(person_list, problems_vect)
}
\arguments{
\item{person_list}{the yaml front matter part containing person info}

\item{problems_vect}{character vector of previously encountered problems}
}
\value{
a character vector of previously encountered problems and problems
identified for person names and \code{orcid} ids.
}
\description{
Checks the format of person names and \code{orcid} ids.
}
\seealso{
Other check:
\code{\link[=check_all]{check_all()}},
\code{\link[=check_frontmatter]{check_frontmatter()}},
\code{\link[=check_structure]{check_structure()}},
\code{\link{protocolcheck}},
\code{\link[=validate_orcid]{validate_orcid()}}
}
\concept{check}
