% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_frontmatter.R
\name{check_frontmatter}
\alias{check_frontmatter}
\title{Checks protocol metadata}
\usage{
check_frontmatter(protocol_code, fail = !interactive())
}
\arguments{
\item{protocol_code}{Character string giving the protocol code}

\item{fail}{Should the function drop an error in case of a problem?
Defaults to \code{TRUE} in a non-interactive session and \code{FALSE} in an interactive
session.}
}
\value{
A report of all failed checks.
}
\description{
This function reads metadata from the yaml front matter stored
in the \code{index.Rmd} file of a protocol and checks if the metadata format is
correct.
This function is intended for checking if a protocol is ready to be rendered
and published (for instance, it will fail if version number is
\code{YYYY.NN.dev}).
}
\seealso{
Other check:
\code{\link[=check_all]{check_all()}},
\code{\link[=check_all_person_info]{check_all_person_info()}},
\code{\link[=check_structure]{check_structure()}},
\code{\link{protocolcheck}},
\code{\link[=validate_orcid]{validate_orcid()}}
}
\concept{check}
