% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{check_protocolcode}
\alias{check_protocolcode}
\title{Check the \code{protocolcode} format}
\usage{
check_protocolcode(protocolcode)
}
\arguments{
\item{version_number}{Character string with format \code{YYYY.NN}}
}
\description{
Check if \code{protocolcode} is of format \verb{s[f|p|i|o|a]p-###-[nl|en]}
}
\keyword{internal}
