% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{check_structure}
\alias{check_structure}
\title{Checks protocol document structure}
\usage{
check_structure(protocol_code, fail = !interactive())
}
\arguments{
\item{protocol_code}{Character string giving the protocol code}

\item{fail}{Should the function drop an error in case of a problem?
Defaults to \code{TRUE} in a non-interactive session and \code{FALSE} in an interactive
session.}
}
\value{
A report of all failed checks.
}
\description{
This function reads the protocol and checks if the document
structure is correct: chunks have head and tail, required headings are
present and in the right order,...
This function is intended for checking if a protocol is ready to be rendered
and published.
}
\seealso{
Other check:
\code{\link[=check_all]{check_all()}},
\code{\link[=check_all_person_info]{check_all_person_info()}},
\code{\link[=check_frontmatter]{check_frontmatter()}},
\code{\link{protocolcheck}},
\code{\link[=validate_orcid]{validate_orcid()}}
}
\concept{check}
