% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_protocolsection.R
\name{check_versionnumber}
\alias{check_versionnumber}
\title{Check the version number format}
\usage{
check_versionnumber(version_number)
}
\arguments{
\item{version_number}{Character string with format YYYY.NN#'}
}
\description{
Check if version number is of format YYYY.NN
}
\keyword{internal}
