% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{conditional_return_on_error}
\alias{conditional_return_on_error}
\title{Early return upon conditional error check}
\usage{
conditional_return_on_error(condition, message, checkobject)
}
\arguments{
\item{condition}{An R expression resulting in TRUE or FALSE.}

\item{message}{A string containing the error message when condition is TRUE.}

\item{checkobject}{A protocolhelper check object.}
}
\value{
A protocol check object
}
\description{
If a condition is met, add an error message to the
protocolhelper check object and return it.
}
\keyword{internal}
