% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_docx.R
\name{convert_docx_to_rmd}
\alias{convert_docx_to_rmd}
\title{Convert a docx-file (a protocol) to an (R)markdown file}
\usage{
convert_docx_to_rmd(
  from,
  to = sub("docx$", "Rmd", from),
  dir_media = ".",
  wrap = NA,
  overwrite = FALSE,
  verbose = FALSE,
  wd = getwd()
)
}
\arguments{
\item{from}{The \code{.docx} file to convert.
Can be given as an absolute or relative path.}

\item{to}{The filename including path to write the resulting \code{.Rmd} file.
The default is to use the same name and path as the \code{.docx} document.}

\item{dir_media}{The directory to write the folder \code{media} with images to,
relative to the folder where the \code{.Rmd} is written.
Defaults to '.' (the path where the \code{.Rmd} file is written).}

\item{wrap}{The width at which to wrap text.
If \code{NA} (default), text is not wrapped.}

\item{overwrite}{Whether or not to overwrite the \code{to} file if it already
existed.
Defaults to \code{FALSE}.}

\item{verbose}{Whether to print \code{pandoc} progress text.
Defaults to \code{FALSE}.}

\item{wd}{Current working directory (used to handle relative paths).}
}
\description{
This function is derived from the
\code{redoc::dedoc()}
function and uses \code{pandoc} to convert between docx and markdown.
Several options are preset to end-up with a markdown document that is in
syntax as close as possible to Rmarkdown files in RStudio.
During conversion, graphics (e.g. png, jpg) will be extracted from the docx
archive and placed in a folder \code{./media} and named \code{image1}, \code{image2},
etcetera.
Additionally, .emf files will be converted to .png.
}
\details{
Metadata in the page headers and footers of the docx are ignored
and will thus be lost during conversion.
In case the header or footer did contain important metadata, it will need to
be recovered manually.
Usually header information will go inside a yaml section of an Rmarkdown.
}
\seealso{
Other convert:
\code{\link[=add_captions]{add_captions()}}
}
\concept{convert}
