% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_docx.R
\name{docx_to_md}
\alias{docx_to_md}
\title{Convert a docx-file (a protocol) to an (R)markdown file}
\usage{
docx_to_md(
  docx,
  to = NULL,
  dir = ".",
  wrap = 80,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{docx}{The \code{.docx} file to convert. Can be given as an absolute or relative path.}

\item{to}{The filename to write the resulting \code{.Rmd} file (without path). The default is to
use the same basename as the docx document.}

\item{dir}{The directory to write the \code{.Rmd} to. Defaults to current working
directory. Any images will be written to \code{dir/media}.}

\item{wrap}{The width at which to wrap text. If \code{NA}, text is not wrapped.
Defaults to 80.}

\item{overwrite}{Whether or not to overwrite the to file if it already
existed. Defaults to FALSE.}

\item{verbose}{whether to print pandoc progress text. Defaults to FALSE.}
}
\description{
This function is derived from the \link[redoc]{dedoc} and uses pandoc to convert between docx and markdown. Several options are preset to end-up with a markdown document that is in syntax as close as possible to Rmarkdown files in RStudio. During conversion, graphics (e.g. png, jpg) will be extracted from the docx archive and placed in a folder \code{./media} and named \code{image1}, \code{image2}, etcetera.
}
\details{
Metadata in the page headers and footers of the docx are ignored and will thus be lost during conversion. In case the header or footer did contain important metadata, it will need to be recovered manually. Usually header information will go inside a yaml section of an Rmarkdown.
}
