% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{execshell}
\alias{execshell}
\title{Pass command lines to a shell}
\usage{
execshell(commandstring, intern = FALSE, ...)
}
\arguments{
\item{commandstring}{String.
One or multiple commandlines as one would enter in the shell.}

\item{intern}{a logical (not \code{NA}) which indicates whether to
    capture the output of the command as an \R character vector.}

\item{...}{Other arguments passed to \code{\link[base]{system}} or
\code{\link[base]{shell}}.}
}
\description{
Cross-platform function to pass a command to the shell, using either
\code{\link[base]{system}} or
(Windows-only) \code{\link[base]{shell}}, depending on the operating system.
}
\keyword{internal}
