% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fence_all_chunks.R
\name{fence_all_chunks}
\alias{fence_all_chunks}
\title{Surround chunks with \code{pandoc} markdown \code{divs}}
\usage{
fence_all_chunks(input)
}
\arguments{
\item{input}{Path to input \code{Rmarkdown} file}
}
\description{
Adapted from \code{bookdown::fence_theorems()} but with the following differences:
\itemize{
\item chunks will be kept, not replaced by \code{divs}
\item all chunks will be processed, not only theorem chunks
\item all chunks will be surrounded by \code{div} environments
\item a label will be added to the \code{div} environment based on the chunk name
\item if the chunk parameters contain \code{fig.cap}, the label will be
\verb{#fig:<chunkname>}, else we assume it is a table \verb{#tab:<chunkname>}
}
}
\details{
This function is only relevant in case of
\code{bookdown::markdown_document2} output in order to work around a problem with
cross-references for tables and figures.
It is assumed that the \verb{\\@ref()} syntax is used to refer to tables and
figures and that chunk names are used as labels to refer to figures and
tables.
The input file will be overwritten.
}
\keyword{internal}
