% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_protocolsection.R
\name{get_data_media}
\alias{get_data_media}
\title{Internal function to get media and data files.

Extracts all paths to files in data and media from rmd_content and uses git
show to copy paste the files from the subprotocol to the main protocol}
\usage{
get_data_media(rmd_content, code_subprotocol, tag)
}
\arguments{
\item{code_subprotocol}{Character string giving the protocol code from
which a section or sections will be extracted}

\item{tag}{}
}
\value{
Silent.
}
\description{
Internal function to get media and data files.

Extracts all paths to files in data and media from rmd_content and uses git
show to copy paste the files from the subprotocol to the main protocol
}
\keyword{internal}
