% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get protocol tags and optionally bump version number}
\usage{
get_tags(protocol_code, bump_version = FALSE)
}
\arguments{
\item{protocol_code}{Protocol code for which you want to get the tags}

\item{bump_version}{whether or not to update the development version number
to the incremented version number in the YAML header of the index.Rmd file
of the target protocol}
}
\value{
A message containing both the general and specific tag
}
\description{
Searches all index.Rmd files for version numbers and constructs tags based
on the protocol-code and an incremented version number.
The function assumes that you have the development branch checked out and
that this branch is aligned with the main branch.
}
