% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_protocolcheck.R
\name{load_protocolcheck}
\alias{load_protocolcheck}
\title{Load protocol to be checked}
\usage{
load_protocolcheck(x)
}
\arguments{
\item{x}{Either a \code{Protocolcheck} object or the variable \code{protocol_code}.}
}
\value{
A \code{Protocolcheck} object.
}
\description{
Make a \verb{Protocolcheck object} for the protocol that has to be
checked if the argument is a protocol code.
}
