% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_protocolsection.R
\name{mark_chunks}
\alias{mark_chunks}
\title{Mark line as part of chunk}
\usage{
mark_chunks(rmd_content)
}
\arguments{
\item{rmd_content}{Character vector (each element is a line read from an
Rmarkdown file)}
}
\value{
A logical vector
}
\description{
The function will search for beginning and end of chunk and mark all lines
in between as TRUE, otherwise FALSE
}
\keyword{internal}
