% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{protocol_tex}
\alias{protocol_tex}
\title{set path to \code{tex} template to be used by \code{pdf_book}}
\usage{
protocol_tex()
}
\description{
set path to \code{tex} template to be used by \code{pdf_book}
}
\keyword{internal}
