% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all.R
\name{render_all}
\alias{render_all}
\title{Function to render all protocols into a folder
\code{publish/version_number}.}
\usage{
render_all(output_root = "publish", version_number)
}
\arguments{
\item{output_root}{A character string giving the root folder without version
number. Default is "publish".}

\item{version_number}{A character string of the form \code{YYYY.NN} (year, number
giving order of release in a year).}
}
\description{
This function renders all protocols (which have been merged to
the master branch) into a folder \code{publish/version_number}, where
\code{version_number} is of the form \code{YYYY.NN} (year, number giving order of
release in a year).
}
\details{
The rendered html files and associated files needed by the html file
for all protocols will be written to subfolders of \code{publish/version_number}.
The subfolder structure mirrors the structure of \code{src/thematic} and
\code{src/project}.
}
\examples{
\dontrun{
render_all(version_number = "2000.01")
}
}
