% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_release.R
\name{render_release}
\alias{render_release}
\title{Function to render the protocols that are ready for a new release.}
\usage{
render_release(output_root = "publish")
}
\arguments{
\item{output_root}{A character string giving the root folder without version
number. Default is "publish".}
}
\description{
This function renders the protocols which are ready for a new
release into a folder \code{publish/version_number/language}, where
\code{version_number} is of the form \code{YYYY.NN} (year, number giving order of
release in a year) and \code{language} can be \code{nl} or \code{en} (both folders will
be created in case of multi-language protocols). In addition, earlier
released and older versions of protocols remain available in the \code{publish}
folder (see also \code{details}).
}
\details{
The links to earlier releases and older versions can be found
through a subfolder structure below \code{publish/version_number} that mirrors the
structure of \code{src/thematic} and \code{src/project}.
}
\examples{
\dontrun{
protocolhelper:::render_release()
}
}
\keyword{internal}
