% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tags.R
\name{set_tags}
\alias{set_tags}
\title{Set general and specific protocol tags}
\usage{
set_tags(protocol_code, commit = TRUE, path = ".")
}
\arguments{
\item{protocol_code}{The protocol_code corresponding with the name of the
branch that contains the new or updated protocol.}

\item{commit}{Logical. Default TRUE. Whether or not to add and commit the
changes to the protocol branch}

\item{path}{Default is current working directory. Should correspond with
root directory of protocolsource repo.}
}
\value{
invisible NULL
}
\description{
This function is a part of the GitHub Action.
Therefore it only works when run in a GitHub Action on the main or master
branch.
Otherwise it will only return a message.
It sets a general protocol tag at the current commit.
The tag message is generated from the related entry from \code{NEWS.md} as
message.
This tag will turn into a release.
}
