% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_news_release.R
\name{update_news_release}
\alias{update_news_release}
\title{Updates the protocolsource repo NEWS.md file}
\usage{
update_news_release(protocol_code, path = ".")
}
\arguments{
\item{protocol_code}{protocol code of the protocol that is to be published in
this release}

\item{path}{Default current working directory. The root folder of the
protocolsource repo.}
}
\description{
Constructs and writes a NEWS.md file based on version numbers, protocol codes
and protocol titles.
Not intended for interactive use.
Should be used in a GitHub Action.
Assumes the branch containing the protocol for the next release is checked
out, and it is up to date with the main branch.
}
