% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_protocol.R
\name{update_protocol}
\alias{update_protocol}
\title{Preparatory steps to start the update of a pre-existing version of a protocol}
\usage{
update_protocol(protocol_code)
}
\arguments{
\item{protocol_code}{Character string giving the protocol code}
}
\value{
NULL invisibly
}
\description{
The function creates a branch with the same name as the \code{protocol_code} and
checks out that branch.
Next, it will update the version number in the YAML of \code{index.Rmd} and
update the protocol-specific \code{NEWS.md} file.
}
\seealso{
Other creation:
\code{\link[=add_dependencies]{add_dependencies()}},
\code{\link[=add_one_subprotocol]{add_one_subprotocol()}},
\code{\link[=add_subprotocols]{add_subprotocols()}},
\code{\link[=create_protocol]{create_protocol()}},
\code{\link[=insert_protocolsection]{insert_protocolsection()}},
\code{\link[=update_version_number]{update_version_number()}}
}
\concept{creation}
