% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{write_bookdown_yml}
\alias{write_bookdown_yml}
\title{Writes a \verb{_bookdown.yml} file}
\usage{
write_bookdown_yml(
  language,
  book_filename,
  path_to_protocol,
  output_dir_rel,
  rmd_files
)
}
\arguments{
\item{language}{the language of the book}

\item{book_filename}{the filename of the book}

\item{path_to_protocol}{the path to the protocol}

\item{output_dir_rel}{relative output directory}

\item{rmd_files}{character vector of R Markdown filenames}
}
\description{
Creates contents from its arguments and writes to file \verb{_bookdown.yml}
}
\keyword{internal}
