% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{write_output_yml}
\alias{write_output_yml}
\title{Writes \verb{_output.yml} file}
\usage{
write_output_yml(language, path_to_protocol)
}
\arguments{
\item{language}{language of the protocol}

\item{path_to_protocol}{path to the protocol}
}
\description{
Creates and writes a \verb{_output.yml} file
}
\keyword{internal}
