% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{write_yaml_front_matter}
\alias{write_yaml_front_matter}
\title{Fills in values from key-value pairs in \code{yaml} front matter}
\usage{
write_yaml_front_matter(
  parent_rmd,
  path_to_protocol,
  title,
  subtitle,
  date,
  reviewers,
  file_manager,
  version_number,
  protocol_code,
  language,
  authors,
  orcids,
  protocol_type,
  template,
  theme,
  project_name
)
}
\arguments{
\item{parent_rmd}{original \code{index.Rmd} file}

\item{path_to_protocol}{path to the protocol}

\item{title}{A character string giving the main title of the protocol}

\item{subtitle}{A character string for an optional subtitle}

\item{date}{A character string of the date in ISO 8601 format (\code{YYYY-MM-DD})}

\item{reviewers}{A character vector for reviewers of the form First name
Last name}

\item{file_manager}{A character string for the name of the document
maintainer of the form First name Last name}

\item{version_number}{A version number of the form \code{YYYY.##}.
The default is a function which will determine this number automatically.
It should normally not be changed.}

\item{protocol_code}{the protocol code}

\item{language}{Language of the protocol, either \code{"nl"} (Dutch),
the default, or \code{"en"} (English).}

\item{authors}{A character vector for authors of the form
\code{c("lastname1, firstname1", "lastname2, firstname2")}}

\item{orcids}{A character vector of \code{orcid} IDs, equal in length to authors.
If one of the authors does not have an \code{orcid} ID, use \code{NA} to indicate this
in the corresponding position of the character vector (or get an \code{orcid} ID).}

\item{protocol_type}{the protocol type}

\item{template}{Which template to use?
Default is set equal to protocol_type.
However, you can also set this to \code{"generic"} in which case a simplified
template will be used that can be used as an alternative to the default
templates.}

\item{theme}{A character string equal to one of \code{"generic"} (default),
\code{"water"}, \code{"air"}, \code{"soil"}, \code{"vegetation"} or \code{"species"}. It is used as
the folder location (\code{source/sfp/theme}) where standard field protocols
that belong to the same theme will be stored.
Ignored if protocol_type is other than \code{"sfp"}.}

\item{project_name}{A character string that is used as the folder location
(\code{source/spp/project_name}) where project-specific protocols that belong to
the same project will be stored. Preferably a short name or acronym. If the
folder does not exist, it will be created.
Ignored if protocol_type is other than \code{"spp"}.}
}
\description{
Overwrites the \code{index.Rmd} file
}
\keyword{internal}
