% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dependencies.R
\name{add_dependencies}
\alias{add_dependencies}
\title{Adds dependencies to the YAML of an index.Rmd file}
\usage{
add_dependencies(
  code_mainprotocol,
  protocol_code,
  version_number,
  params,
  appendix = !is.na(params)
)
}
\arguments{
\item{code_mainprotocol}{Protocol code of the protocol for which dependencies
need to be declared in the YAML of its index.Rmd file}

\item{protocol_code}{Character vector of protocol codes that are dependencies
to the main protocol.}

\item{version_number}{Character vector of version numbers corresponding with
protocol_code.}

\item{params}{List of lists with protocol-specific parameters corresponding
with parameters from the protocols in protocol_code. Use \code{NA} if no
parameters should be set for a protocol.}

\item{appendix}{Logical vector indicating whether or not a dependency needs
to be included as a subprotocol (at the end of the main protocol in an
appendix). Default is \code{!is.na(params)}. When \code{params} is not \code{NA}, \code{appendix}
will always be set to \code{TRUE} even if the user passes another value.}
}
\description{
Adds dependencies to the YAML of an index.Rmd file
}
