% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_label.R
\name{add_label}
\alias{add_label}
\title{Helper function to create labelled captions for \code{pander} tables}
\usage{
add_label(caption = "", tag = "tab")
}
\arguments{
\item{caption}{The caption text as a string}

\item{tag}{The tag to use as a prefix. Default is \code{tab}.}
}
\value{
The caption text prefixed with a reference label.
}
\description{
The function adds a reference label to the caption so that \code{pander} tables
can be cross-referenced in \code{bookdown} using the \verb{\\@ref(tab:label)} syntax.
The function should only be used in \code{pander(x, caption = add_label())}.
}
\examples{
add_label("caption text")
}
\seealso{
Other utility: 
\code{\link{get_path_to_protocol}()},
\code{\link{get_protocol_type}()},
\code{\link{get_protocolnumbers}()},
\code{\link{get_short_titles}()},
\code{\link{get_version_number}()},
\code{\link{increment_version_number}()}
}
\concept{utility}
