% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_subprotocols.R
\name{add_one_subprotocol}
\alias{add_one_subprotocol}
\title{Helper function to add one subprotocol to a project-specific protocol
of which it is a dependency}
\usage{
add_one_subprotocol(
  code_subprotocol,
  version_number,
  params2 = NULL,
  code_mainprotocol,
  fetch_remote = TRUE
)
}
\arguments{
\item{code_subprotocol}{Character string giving the protocol code from
which a subprotocol will be made (usually a sfp-type protocol)}

\item{version_number}{Character string with format YYYY.NN}

\item{params2}{A list of parameter key-value pairs.}

\item{code_mainprotocol}{Character string giving the protocol code for the
main protocol}

\item{fetch_remote}{Whether or not to fetch the remote. Default TRUE.}
}
\description{
The function renders the subprotocol to
\code{\link[bookdown:html_document2]{bookdown::markdown_document2()}} and
saves the resulting md file (and any associated media and data files) in a
subfolder of the directory of the project-specific protocol.
This function should normally not be called directly.
Use \code{\link[=add_subprotocols]{add_subprotocols()}} instead.
}
\examples{
\dontrun{
  add_subprotocols(code_mainprotocol = 'spp-999-en')
}
}
