% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_subprotocols.R
\name{add_subprotocols}
\alias{add_subprotocols}
\title{Render all subprotocols to single markdown files}
\usage{
add_subprotocols(.dependencies, ...)
}
\arguments{
\item{.dependencies}{a data.frame with columns version_number, params and
appendix. Each column corresponds with an argument of
\code{\link[=add_one_subprotocol]{add_one_subprotocol()}}}

\item{...}{additional parameters passed on to
\code{\link[=add_one_subprotocol]{add_one_subprotocol()}}.}
}
\value{
For each subprotocol a single markdown file and associated media and
data files. Each subprotocol will be written to a subfolder of the main
protocol. The subfolder name is the same as the version number of the
subprotocol.
}
\description{
The function should be called in a separate R script.
}
