% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_frontmatter.R
\name{check_frontmatter}
\alias{check_frontmatter}
\title{Checks protocol metadata}
\usage{
check_frontmatter(protocol_folder_name)
}
\arguments{
\item{protocol_folder_name}{Character string giving the name of the protocol subfolder}
}
\value{
Silent, or an error message stating which check failed.
}
\description{
This function reads metadata from the yaml front matter stored
the index.Rmd file of a protocol and checks if the metadata format is correct.
This function is intended for checking if a protocol is ready to be rendered
and published (for instance, it will fail if version number is \code{YYYY.NN.dev}).
}
