% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_frontmatter.R
\name{check_frontmatter}
\alias{check_frontmatter}
\title{Checks protocol metadata}
\usage{
check_frontmatter(protocol_code, yaml = TRUE)
}
\arguments{
\item{protocol_code}{Character string giving the protocol code}

\item{yaml}{Boolean. If \code{FALSE} the function will only print error
messages, otherwise (default) the parsed yml front matter will be returned}
}
\value{
If one of the checks fails, an error message will be returned.
Otherwise, the parsed yml front matter if \code{yaml} is
\code{TRUE} is returned, or a message that everything is OK otherwise.
}
\description{
This function reads metadata from the yaml front matter stored
the index.Rmd file of a protocol and checks if the metadata format is
correct.
This function is intended for checking if a protocol is ready to be rendered
and published (for instance, it will fail if version number is
\code{YYYY.NN.dev}).
}
