% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_protocolsection.R
\name{check_protocolcode}
\alias{check_protocolcode}
\title{Check the protocolcode format}
\usage{
check_protocolcode(protocolcode)
}
\arguments{
\item{version_number}{Character string with format YYYY.NN#'}
}
\description{
Check if protocolcode is of format \verb{s[f|p|i|o|a]p-###-[nl|en]}
}
\keyword{internal}
