% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{check_structure}
\alias{check_structure}
\title{Checks protocol document structure}
\usage{
check_structure(protocol_code)
}
\arguments{
\item{protocol_code}{Character string giving the protocol code}
}
\value{
If one of the checks fails, an error message will be returned.
}
\description{
This function reads the protocol and checks if the document
structure is correct: chunks have head and tail, required headings are
present and in the right order,...
This function is intended for checking if a protocol is ready to be rendered
and published.
}
