% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_docx.R
\name{convert_docx_to_rmd}
\alias{convert_docx_to_rmd}
\title{Convert a docx-file (a protocol) to an (R)markdown file}
\usage{
convert_docx_to_rmd(
  from,
  to,
  dir,
  wrap = 80,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{from}{The \code{.docx} file to convert.
Can be given as an absolute or relative path.}

\item{to}{The filename to write the resulting \code{.Rmd} file (without path).
The default is to use the same basename as the \code{.docx} document.}

\item{dir}{The directory to write the \code{.Rmd} to.
Defaults to current working directory.
Any images will be written to \code{dir/media}.}

\item{wrap}{The width at which to wrap text.
If \code{NA}, text is not wrapped.
Defaults to 80.}

\item{overwrite}{Whether or not to overwrite the \code{to} file if it already
existed.
Defaults to \code{FALSE}.}

\item{verbose}{Whether to print pandoc progress text.
Defaults to \code{FALSE}.}
}
\description{
This function is derived from the
\href{https://noamross.github.io/redoc/reference/dedoc.html}{\code{redoc::dedoc()}}
function and uses pandoc to convert between docx and markdown.
Several options are preset to end-up with a markdown document that is in
syntax as close as possible to Rmarkdown files in RStudio.
During conversion, graphics (e.g. png, jpg) will be extracted from the docx
archive and placed in a folder \code{./media} and named \code{image1}, \code{image2},
etcetera.
}
\details{
Metadata in the page headers and footers of the docx are ignored
and will thus be lost during conversion.
In case the header or footer did contain important metadata, it will need to
be recovered manually.
Usually header information will go inside a yaml section of an Rmarkdown.
}
