% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol_helpers.R
\name{create_from_docx}
\alias{create_from_docx}
\title{Create an Rmarkdown version from an existing \code{docx} protocol}
\usage{
create_from_docx(from_docx, path_to_protocol)
}
\arguments{
\item{from_docx}{A character string with the path (absolute or relative) to
a \code{.docx} file containing a pre-existing protocol.}

\item{path_to_protocol}{Absolute path to the protocol folder where the
protocol created from \code{docx} needs to be written to}
}
\description{
The \code{docx} file is first converted to a single \code{Rmd} file with the aid of
\code{pandoc} (called from \code{convert_docx_to_rmd}).
Any emf images are converted to png.
Next, the file is split by chapter in multiple \code{Rmd} files.
All graphics files will be stored in a ./media folder.
Bookdown compatible captions and cross-references for Figures and Tables are
added if and only if \code{'Figuur'} and \code{'Tabel'} is used in the original
document.
}
\keyword{internal}
