% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{create_protocol_code}
\alias{create_protocol_code}
\title{Create protocol code from it's components}
\usage{
create_protocol_code(protocol_type, theme, protocol_number, language)
}
\arguments{
\item{protocol_type}{Either \code{sfp} (standard field protocol) or \code{spp} (
standard project protocol)}

\item{theme}{A character string equal to one of \code{"generic"} (default),
\code{"water"}, \code{"air"}, \code{"soil"}, \code{"vegetation"} or \code{"species"}. It is used as
the folder location (\code{src/thematic/theme}) where standard field protocols
that belong to the same theme will be stored.
Ignored if protocol_type = \code{"spp"}.}

\item{protocol_number}{A character string giving the protocol number.
This parameter should normally not be specified (i.e. NULL), unless
\code{from_docx} is specified.
A protocol number is a three digit string where the first digit corresponds
with a theme and the last two digits identify a protocol within a theme for
standard field protocols. A protocol number for a project-specific protocol
is just a three digit string.
If NULL (the default), a protocol number will be determined automatically
based on pre-existing protocol numbers.
Protocol numbers that are already in use can be retrieved with
\code{get_protocolnumbers()}.}

\item{language}{Language of the protocol, either \code{"nl"} (Dutch),
the default, or \code{"en"} (English).}
}
\value{
A character string containing the protocol_code
}
\description{
A protocol code of format \verb{s[f|p]p-###-[nl|en]} will be created.
The number will be determined automatically based on theme (in case of sfp)
and a rank order of all existing
}
\keyword{internal}
