% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol_helpers.R
\name{create_protocol_code}
\alias{create_protocol_code}
\title{Create protocol code from it's components}
\usage{
create_protocol_code(protocol_type, theme, protocol_number, language)
}
\arguments{
\item{protocol_type}{Either \code{sfp} (standard field protocol), \code{spp} (
standard project protocol), \code{sap} (standard analytical protocol), \code{sip} (
standard instrument protocol), \code{sop} (standard operating protocol)}

\item{theme}{A character string equal to one of \code{"generic"} (default),
\code{"water"}, \code{"air"}, \code{"soil"}, \code{"vegetation"} or \code{"species"}. It is used as
the folder location (\code{source/sfp/theme}) where standard field protocols
that belong to the same theme will be stored.
Ignored if protocol_type is other than \code{"sfp"}.}

\item{protocol_number}{A character string giving the protocol number.
This parameter should normally not be specified (i.e. NULL), unless
\code{from_docx} is specified.
A protocol number is a three digit string where the first digit corresponds
with a theme and the last two digits identify a protocol within a theme for
standard field protocols. A protocol number for other protocol types
is just a three digit string.
If NULL (the default), a protocol number will be determined automatically
based on pre-existing protocol numbers.
Note that for backwards compatibility with protocol numbers that were already
in use at INBO, we made a list of reserved numbers.
These reserved numbers will not be used when \code{protocol_number} is NULL.
The only time you will need to explicitly pass a protocol number to the
\code{protocol_number} argument is when you want to migrate a pre-existing INBO
protocol to \code{protocolsource} and hence use one of the reserved numbers.
Protocol numbers that are already in use in \code{protocolsource} can be retrieved
with \code{get_protocolnumbers()}.}

\item{language}{Language of the protocol, either \code{"nl"} (Dutch),
the default, or \code{"en"} (English).}
}
\value{
A character string containing the protocol_code
}
\description{
A protocol code of format \verb{s[fpioa]p-###-[nl|en]} will be created.
The number will be determined automatically based on theme (in case of \code{sfp})
and a rank order of all existing and reserved protocol numbers, unless
the protocol number is passed directly to the \code{protocol_number} argument.
}
\keyword{internal}
