% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{create_sfp}
\alias{create_sfp}
\title{Create a folder with a markdown template to start working on a new fieldwork protocol}
\usage{
create_sfp(
  title = "Titel",
  subtitle = "",
  short_title = "korte titel",
  authors = "voornaam1 naam1, voornaam2 naam2, ...",
  date = Sys.Date(),
  reviewers = "voornaam1 naam1, voornaam2 naam2, ...",
  file_manager = "voornaam naam",
  revision = "0.0.0.9000",
  projectname = "",
  theme = c("generiek", "water", "lucht", "bodem", "vegetatie", "soorten")
)
}
\arguments{
\item{title}{A character string giving the main title of the protocol}

\item{subtitle}{A character string for an optional subtitle}

\item{short_title}{A character string of less than 20 characters to use in folder and filenames}

\item{authors}{A character string for authors of the form Voornaam Achternaam and multiple authors separated by a comma}

\item{date}{A character string of the date in ISO 8601 format (YYYY-MM-DD)}

\item{reviewers}{A character string for reviewers of the form Voornaam Achternaam and multiple authors separated by a comma}

\item{file_manager}{A character string for the name of the document maintainer}

\item{revision}{A semantic version number of the form major.minor.patch. For development versions a fourth component is appended starting from .9000. The default is 0.0.0.9000 and should normally not be changed.}

\item{projectname}{A character string giving the project name. Best to use an acronym for this.}

\item{theme}{A character string equal to generiek (default), water, lucht, bodem, vegetatie of soorten.}
}
\value{
A new subfolder beneath src with an Rmarkdown file
}
\description{
This function will create a new folder based on values that are passed on via the parameters and creates a R-markdown skeleton based on a template file to start working on a new protocol.
}
\examples{
\dontrun{
create_sfp()
}
}
