% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{create_sfp}
\alias{create_sfp}
\title{Create a folder with a bookdown (R markdown) template to start a new
fieldwork protocol and optionally render to html}
\usage{
create_sfp(
  title,
  subtitle,
  short_title,
  authors,
  date = Sys.Date(),
  reviewers,
  file_manager,
  version_number = "YYYY.##.dev",
  theme = c("generic", "water", "air", "soil", "vegetation", "species"),
  language = c("nl", "en"),
  from_docx = NULL,
  protocol_number = NULL,
  render = FALSE
)
}
\arguments{
\item{title}{A character string giving the main title of the protocol}

\item{subtitle}{A character string for an optional subtitle}

\item{short_title}{A character string of less than 20 characters to use in
folder and filenames}

\item{authors}{A character vector for authors of the form First name Last
name}

\item{date}{A character string of the date in ISO 8601 format (YYYY-MM-DD)}

\item{reviewers}{A character vector for reviewers of the form First name
Last name}

\item{file_manager}{A character string for the name of the document
maintainer of the form First name Last name}

\item{version_number}{A version number of the form \code{YYYY.##}.
For development versions \code{.dev} is added.
The default is \code{YYYY.##.dev} (See \code{from_docx}).
When the protocol is ready to be released, this should be changed by a repo-
admin.}

\item{theme}{A character string equal to one of \code{"generic"} (default),
\code{"water"}, \code{"air"}, \code{"soil"}, \code{"vegetation"} or \code{"species"}.}

\item{language}{Language of the protocol, either \code{"nl"} (Dutch),
the default, or \code{"en"} (English).}

\item{from_docx}{A character string with the path (absolute or relative) to
a \code{.docx} file containing a pre-existing protocol.
Please make sure to copy-paste all relevant meta-data from the \code{.docx} file
to the corresponding parameters of this function.
If nothing is provided (i.e. default = NULL), an empty template will be used.}

\item{protocol_number}{A character string giving the protocol number.
This parameter should normally not be specified (i.e. NULL), unless
\code{from_docx} is specified.
A protocol number is a three digit string where the first digit corresponds
with a theme and the last two digits identify a protocol within a theme.
If NULL (the default), a protocol number will be determined automatically
based on pre-existing protocol numbers.
Protocol numbers that are already in use can be retrieved with
\code{get_protocolnumbers()}.}

\item{render}{Whether or not to render the protocol to html.
Defaults to FALSE.}
}
\value{
A target folder to which files will be written will be created as
subdirectories beneath \code{src}.
The subfolder structure is of the form
\verb{/thematic/theme/sfp-protocolnumber_short_title_language/} where \code{theme},
\code{protocolnumber}, \code{short_title} and \code{language} are determined by the
corresponding arguments of the function.
A matching subfolder structure will be created beneath the \code{docs} folder (and
output files needed for rendering to html output will be placed in it if
\code{render = TRUE}.
The template Rmarkdown files, or the Rmarkdown files that result from
converting a docx protocol (see \code{from_docx} argument), will be written to
the target folder beneath \code{src}.
Besides Rmarkdown files, this target folder will also contain files needed to
render to a Bookdown gitbook such as a \verb{_bookdown.yml}.
The \verb{00_NEWS.Rmd} file must be used to document the changes between revisions
of the protocol.
Furthermore, a \code{data} and a \code{media} folder will be created as subdirectories
of the target folder.
The \code{media} folder will contain image files extracted from the docx protocol
when the \code{from_docx} argument is used.
The \code{data} folder can be used to store tabular data that are needed for the
protocol.
}
\description{
This function will create a new folder based on values that are
passed on via the parameters and creates a R-markdown (bookdown) skeleton
based on a template file to start working on a new protocol.
Optionally, the rmarkdown chapters are rendered to an html files which will
be saved in a matching subfolder of the \code{docs} folder.
}
\details{
It is assumed that the \code{src} folder is a subfolder of an RStudio
project with git version control.
}
