% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{create_sfp}
\alias{create_sfp}
\title{Create a folder with a bookdown (R markdown) template to start a new fieldwork protocol}
\usage{
create_sfp(
  title,
  subtitle,
  short_title,
  authors,
  date = Sys.Date(),
  reviewers,
  file_manager,
  revision = "0.0.0.9000",
  theme = c("generic", "water", "air", "soil", "vegetation", "species"),
  language = c("nl", "en")
)
}
\arguments{
\item{title}{A character string giving the main title of the protocol}

\item{subtitle}{A character string for an optional subtitle}

\item{short_title}{A character string of less than 20 characters to use in folder and filenames}

\item{authors}{A character string for authors of the form First name Last name and multiple authors separated by a comma}

\item{date}{A character string of the date in ISO 8601 format (YYYY-MM-DD)}

\item{reviewers}{A character string for reviewers of the form First name Last name and multiple authors separated by a comma}

\item{file_manager}{A character string for the name of the document maintainer of the form First name Last name}

\item{revision}{A semantic version number of the form major.minor.patch. For development versions a fourth component is appended starting from .9000. The default is 0.0.0.9000 and should normally not be changed.}

\item{theme}{A character string equal to one of \code{"generic"} (default),
\code{"water"}, \code{"air"}, \code{"soil"}, \code{"vegetation"} or \code{"species"}.}

\item{language}{Language of the protocol, either \code{"nl"} (Dutch), the default, or \code{"en"} (English).}
}
\description{
This function will create a new folder based on values that are passed on via the parameters and creates a R-markdown (bookdown) skeleton based on a template file to start working on a new protocol.
}
\details{
The created folder is a new subfolder beneath \code{src} and contains a bookdown project. It is assumed that the src folder is a subfolder of an RStudio project with git version control.
}
\examples{
\dontrun{
create_sfp()
}
}
