% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_to_protocol.R
\name{get_path_to_protocol}
\alias{get_path_to_protocol}
\title{Function to get (or set) the full path to a protocol}
\usage{
get_path_to_protocol(
  protocol_code,
  theme = NULL,
  project_name = NULL,
  short_title = NULL
)
}
\arguments{
\item{protocol_code}{Character string giving the protocol code}

\item{theme}{A character string equal to one of \code{"generic"},
\code{"water"}, \code{"air"}, \code{"soil"}, \code{"vegetation"} or \code{"species"}.
Defaults to NULL.
Only needed if no folder with the name of the protocol code exists and
the request is for a sfp protocol.}

\item{project_name}{Character string giving the name of the project folder.
Defaults to NULL.
Only needed if no folder with the name of the protocol code exists and
the request is for a spp protocol.}

\item{short_title}{A character string of less than 20 characters to use in
folder and filenames.
Defaults to NULL.
Only needed if no folder with the name of the protocol code exists.}
}
\value{
A character vector containing the full path to the protocol.
}
\description{
A function that is used by other functions and should normally not be used
directly.

For existing protocol codes, arguments \code{theme} and \code{project_name} are
always ignored.
The function will return the absolute path for that protocol.

For new sfp or spp protocols, also either the \code{theme} or the \code{project_name}
argument and \code{short_title} are required apart from the \code{protocol_code}.
The function will construct the absolute path where the source code for that
new protocol will be written.
}
\examples{
\dontrun{
get_path_to_protocol(protocol_code = "sfp-401-nl")
}
}
