% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{get_path_to_protocol}
\alias{get_path_to_protocol}
\title{Internal function to get (or set) the full path to a protocol}
\usage{
get_path_to_protocol(
  protocol_code,
  theme = NULL,
  project_name = NULL,
  short_title = NULL
)
}
\arguments{
\item{protocol_code}{Character string giving the protocol code}

\item{theme}{A character string equal to one of \code{"generic"},
\code{"water"}, \code{"air"}, \code{"soil"}, \code{"vegetation"} or \code{"species"}.
Defaults to NULL.
Only needed if no folder with the name of the protocol code exists and
the request is for a thematic protocol.}

\item{project_name}{Character string giving the name of the project folder.
Defaults to NULL.
Only needed if no folder with the name of the protocol code exists and
the request is for a project-specific protocol.}

\item{short_title}{A character string of less than 20 characters to use in
folder and filenames.
Defaults to NULL.
Only needed if no folder with the name of the protocol code exists.}
}
\value{
A character vector containing the full path to the protocol.
}
\description{
Either provide the full path of an existing, given folder name,
or construct one for non-existing protocols using (new) folder name plus
the theme or project name.

For existing protocol folders, arguments \code{theme} and \code{project_name} are
always ignored.

To create a new protocol folder,
also either the \code{theme} or the \code{project_name} argument are required apart
from the \code{protocol_folder_name}.
}
\examples{
\dontrun{
get_path_to_protocol(protocol_code = "sfp-401-nl")
}
}
\keyword{internal}
