% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{get_path_to_protocol}
\alias{get_path_to_protocol}
\title{Internal function to get (or set) the full path to a protocol.

Either provide the full path of an existing, given folder name,
or construct one for non-existing protocols using (new) folder name plus
the theme or project name.

For existing protocol folders, arguments \code{theme} and \code{project} are always
ignored.

To create a new protocol folder,
also either the \code{theme} or the \code{project} argument are required apart from
the \code{protocol_folder_name}.}
\usage{
get_path_to_protocol(protocol_folder_name = NULL, theme = NULL, project = NULL)
}
\arguments{
\item{protocol_folder_name}{Character string giving the name of the protocol
subfolder}

\item{theme}{A character string equal to one of \code{"generic"},
\code{"water"}, \code{"air"}, \code{"soil"}, \code{"vegetation"} or \code{"species"}.
Defaults to NULL.
Only needed if no folder with the name of the protocol subfolder exists and
the request is for a thematic protocol.}

\item{project}{Character string giving the name of the project folder.
Defaults to NULL.
Only needed if no folder with the name of the protocol subfolder exists and
the request if for a project-specific protocol.}
}
\value{
A character vector containing the full path to the protocol.
}
\description{
Internal function to get (or set) the full path to a protocol.

Either provide the full path of an existing, given folder name,
or construct one for non-existing protocols using (new) folder name plus
the theme or project name.

For existing protocol folders, arguments \code{theme} and \code{project} are always
ignored.

To create a new protocol folder,
also either the \code{theme} or the \code{project} argument are required apart from
the \code{protocol_folder_name}.
}
\examples{
\dontrun{
get_path_to_protocol(protocol_folder_name = "sfp_401-vegopname-terrest_nl")
}
}
\keyword{internal}
