% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protocol_type.R
\name{get_protocol_type}
\alias{get_protocol_type}
\title{Get protocol type from protocol code}
\usage{
get_protocol_type(protocol_code, labels = TRUE, auto_identifier = FALSE)
}
\arguments{
\item{protocol_code}{Character vector giving the protocol code(s)}

\item{labels}{Logical. If \code{TRUE} return full labels, else return just the
three letter abbreviation.}

\item{auto_identifier}{Logical.
If \code{TRUE} returns labels following \href{https://pandoc.org/MANUAL.html#extension-auto_identifiers}{\verb{Pandoc's} auto-identifier}
rules.}
}
\value{
A factor with 5 levels corresponding to \code{sfp}, \code{sip}, \code{sap}, \code{sop}
and \code{spp}. The labels depend on \code{auto_identifier} setting.
}
\description{
The protocol type corresponds to the first 3 letters of the protocol code
}
\seealso{
Other utility: 
\code{\link{add_label}()},
\code{\link{get_path_to_protocol}()},
\code{\link{get_protocolnumbers}()},
\code{\link{get_short_titles}()},
\code{\link{get_version_number}()},
\code{\link{increment_version_number}()}
}
\concept{utility}
