% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol.R
\name{get_protocolnumbers}
\alias{get_protocolnumbers}
\title{Function to list all occupied protocol numbers}
\usage{
get_protocolnumbers(protocol_type = c("sfp", "sip", "sap", "sop"))
}
\arguments{
\item{protocol_type}{A character string equal to sfp (default), sip, sap or
sop.}
}
\value{
A character vector with occupied protocol numbers for a specific
protoocol type
}
\description{
This function will search for protocol numbers in filenames of
Rmarkdown files listed underneath the src folder.
The search will be restricted to files of a given protocol type.
}
\examples{
\dontrun{
get_protocolnumbers()
}
}
