% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol_helpers.R
\name{get_short_titles}
\alias{get_short_titles}
\title{Function to list all short titles that are already in use.}
\usage{
get_short_titles(
  protocol_type = c("sfp", "sip", "sap", "sop", "spp"),
  language = c("nl", "en")
)
}
\arguments{
\item{protocol_type}{A character string equal to \code{sfp} (default), \code{sip},
\code{sap}, \code{sop} or \code{spp}.}

\item{language}{Language of the protocol, either \code{"nl"} (Dutch),
the default, or \code{"en"} (English).}
}
\value{
A character vector with short titles that are in use for a given
protocol type.
}
\description{
This function will search for short titles in filenames of
Rmarkdown files listed underneath the source folder.
The search will be restricted to files of a given protocol type and given
language.
}
\examples{
\dontrun{
get_short_titles()
}
}
\seealso{
Other utility: 
\code{\link{add_label}()},
\code{\link{get_path_to_protocol}()},
\code{\link{get_protocol_type}()},
\code{\link{get_protocolnumbers}()},
\code{\link{get_version_number}()},
\code{\link{increment_version_number}()}
}
\concept{utility}
