% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version_number.R
\name{get_version_number}
\alias{get_version_number}
\title{Get version number for a protocol}
\usage{
get_version_number(path = ".")
}
\arguments{
\item{path}{Defaults to current working directory.
This should correspond with the root directory of the \code{protocolsource} repo.}
}
\value{
A string containing the next (incremented) version number
}
\description{
Looks up pre-existing version numbers of protocols in the main branch and
calculates an incremented (next) version number for the currently checkout
branch containing the created/in development/updated/ready to be released
protocol.
Your local main branch needs to up to date (aligned with remote) for this.
If this is not the case, or other issues are detected regarding a non-clean
local git repository - informative error messages will be given.
}
\seealso{
Other utility: 
\code{\link{add_label}()},
\code{\link{get_path_to_protocol}()},
\code{\link{get_protocol_type}()},
\code{\link{get_protocolnumbers}()},
\code{\link{get_short_titles}()},
\code{\link{increment_version_number}()}
}
\concept{utility}
