% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version_number.R
\name{get_version_number}
\alias{get_version_number}
\title{Get version number for a protocol}
\usage{
get_version_number(path = ".")
}
\arguments{
\item{path}{Defaults to current working directory.
This should correspond with the root directory of the protocolsource repo.}
}
\value{
A string containing the next (incremented) version number
}
\description{
Looks up pre-existing version numbers of protocols in the main branch and
calculates an incremented (next) version number for the currently checkout
branch containing the created/in development/updated/ready to be released
protocol.
}
