% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version_number.R
\name{increment_version_number}
\alias{increment_version_number}
\title{Increment version number}
\usage{
increment_version_number(versions)
}
\arguments{
\item{versions}{a character vector with previously published version numbers}
}
\value{
A string containing the next (incremented) version number
}
\description{
Given a set of published protocol version numbers, calculate the next version
number
}
\seealso{
Other utility: 
\code{\link{get_path_to_protocol}()},
\code{\link{get_protocol_type}()},
\code{\link{get_protocolnumbers}()},
\code{\link{get_short_titles}()},
\code{\link{get_version_number}()}
}
\concept{utility}
