% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocolcheck_class.R
\name{protocolcheck}
\alias{protocolcheck}
\title{The protocolcheck R6 class}
\description{
A class which contains all check results.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Protocolcheck-new}{\code{protocolcheck$new()}}
\item \href{#method-Protocolcheck-add_error}{\code{protocolcheck$add_error()}}
\item \href{#method-Protocolcheck-clone}{\code{protocolcheck$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocolcheck-new"></a>}}
\if{latex}{\out{\hypertarget{method-Protocolcheck-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{protocolcheck$new(protocol_code)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocolcheck-add_error"></a>}}
\if{latex}{\out{\hypertarget{method-Protocolcheck-add_error}{}}}
\subsection{Method \code{add_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{protocolcheck$add_error(msg)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocolcheck-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Protocolcheck-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{protocolcheck$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
