% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_protocol.R
\name{render_protocol}
\alias{render_protocol}
\title{Function to render a protocol to html.}
\usage{
render_protocol(protocol_code = NULL, output_dir = NULL, ...)
}
\arguments{
\item{protocol_code}{Character string giving the protocol code}

\item{output_dir}{The output directory. If \code{NULL}, a field named
\code{output_dir} in the configuration file \file{_bookdown.yml} will be
used (possibly not specified, either, in which case a directory name
\file{_book} will be used).}

\item{...}{Arguments to be passed to
\code{rmarkdown::\link{render}()}. For \code{preview_chapter()}, \code{...}
is passed to \code{render_book()}. See \code{rmarkdown::\link{render}()}
and \href{https://bookdown.org/yihui/bookdown/build-the-book.html}{the
bookdown reference book} for details on how output formatting options are
set from YAML or parameters supplied by the user when calling
\code{render_book()}.}
}
\description{
This function is a simple wrapper around
\code{bookdown::render_book()} and can be used to render a protocol to html
in order to preview updates that have been made.
}
\details{
The rendered html file and associated files needed by the html file
will be put in the directory implied by the output_dir parameter.
}
\examples{
\dontrun{
render_protocol(protocol_code = "sfp_401-nl")
}
}
