% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_release.R
\name{render_release}
\alias{render_release}
\title{Function to render the protocols that are ready for a new release.}
\usage{
render_release(output_root = "publish")
}
\arguments{
\item{output_root}{A character string giving the root folder without version
number. Default is "publish".}
}
\description{
This function renders the protocols which are ready for a new
release into a folder \code{publish/version_number}, where
\code{version_number} is of the form \code{YYYY.NN} (year, number giving order of
release in a year). In addition, earlier
released and older versions of protocols remain available in the \code{publish}
folder (see also \code{details}).
This function is for internal use only.
If you are looking for a function to preview your protocol,
see \code{protocolhelper::render_protocol()}.
}
\details{
The website also contains a welcoming page, a news page and separate
pages per protocol-type with overview tables of all versions of published
protocols.
}
\examples{
\dontrun{
protocolhelper:::render_release()
}
}
\keyword{internal}
