% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_protocol.R
\name{update_protocol}
\alias{update_protocol}
\title{Preparatory steps to start the update of a pre-existing version of a protocol}
\usage{
update_protocol(protocol_code)
}
\arguments{
\item{protocol_code}{Character string giving the protocol code}
}
\value{
NULL invisibly
}
\description{
The function creates a branch named after the protocol_code
}
