% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_version_number.R
\name{update_version_number}
\alias{update_version_number}
\title{Updates the version number in the YAML section of a protocol
\code{index.Rmd} file and in protocol \code{NEWS.md}}
\usage{
update_version_number(protocol_code, commit = TRUE, path = ".")
}
\arguments{
\item{protocol_code}{The protocol_code corresponding with the name of the
branch that contains the new or updated protocol.}

\item{commit}{Logical. Default TRUE. Whether or not to add and commit the
changes to the protocol branch}

\item{path}{Default is current working directory. Should correspond with
root directory of \code{protocolsource} repo.}
}
\value{
TRUE if version number in yaml is updated. FALSE otherwise.
}
\description{
Makes use of \code{get_version_number} to get a new version number
and changes this accordingly in the YAML section of \code{index.Rmd} file and
in \code{NEWS.md}.
}
\seealso{
Other creation: 
\code{\link{add_dependencies}()},
\code{\link{add_one_subprotocol}()},
\code{\link{add_subprotocols}()},
\code{\link{create_protocol}()},
\code{\link{insert_protocolsection}()},
\code{\link{update_protocol}()}
}
\concept{creation}
