% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_version_number.R
\name{update_version_number}
\alias{update_version_number}
\title{Updates the version number in the YAML section of a protocol
index.Rmd file}
\usage{
update_version_number(protocol_code, path = ".")
}
\arguments{
\item{protocol_code}{The protocol_code for which you want to update the
version number.}

\item{path}{Default is current working directory. Should correspond with
root directory of protocolsource repo.}
}
\value{
invisible NULL
}
\description{
Updates the version number in the YAML section of a protocol
index.Rmd file
}
