% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_zenodo.R
\name{update_zenodo}
\alias{update_zenodo}
\title{Update the \code{.zenodo.json} file}
\usage{
update_zenodo(json = ".zenodo.json", write = TRUE)
}
\arguments{
\item{json}{a \code{JSON} string, url or file}

\item{write}{whether to write a \code{.zenodo.json} file (TRUE) or to return a
\code{json}
string}
}
\value{
a \code{.zenodo.json} file (write = TRUE) or a \code{json} string
(write = FALSE)
}
\description{
Updating is done by scanning the \code{index.Rmd} files for author metadata and
adding missing author metadata to the list of creators in the \code{.zenodo.json}.
This internal function should only be used by one of the administrators.
}
\keyword{internal}
