% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_helpers.R
\name{validate_orcid}
\alias{validate_orcid}
\title{validate an ORCID string}
\usage{
validate_orcid(orcid)
}
\arguments{
\item{orcid}{An \code{orcid} in the \code{0000-0000-0000-0000} format}
}
\value{
Logical. TRUE if check digit is correct.
}
\description{
Generates check digit as per ISO 7064 11,2.
The last character in the \verb{ORCID ID} is a checksum.
In accordance with ISO/IEC 7064:2003, MOD 11-2,
this checksum must be "0-9" or "X", a capital letter X which represents
the value 10.
}
\examples{
validate_orcid("0000-0002-6378-6229")
}
\seealso{
Other check: 
\code{\link{check_all}()},
\code{\link{check_all_person_info}()},
\code{\link{check_frontmatter}()},
\code{\link{check_structure}()},
\code{\link{protocolcheck}}
}
\concept{check}
